move app protect waf tests
